implement multi-lingual message and signal descriptions
note that for now the default is to assume that comments which don't
explicitly specify a language are in English. An alternative would be
to check if the dictionary containing the comments contains exactly a
single element and use that instead. The problem with this approach is
that comments for different languages might get mixed up more easily:
Assume a supplier based in China who documents its signals only in
Chinese and a OEM based in Brazil who likes to document its stuff
using Portuguese...