ARXML SystemLoader: Use the new ARXML traversial functions
With this patch, ARXML files are *much* freer to use references and
IMO even the python code's readability improves.

Be aware that I'm not sure where exactly references are allowed by the
AUTOSAR standard or how to obtain that information, so I've enabled
them quite liberally. The rationale is that if AUTOSAR does not allow
references in a given context but we support them anyway, well-formed
files will still load fine.

TODO (?): Adapt the EcuExtractLoader class.