attempt to fix the recently appearing test failures of PRs
the assumption is that these errors are caused because github-CI
recently removed gcc-8 and -- maybe -- clang-8.

While I see the point of using older compilers to ensure that they
stay supported, the code to be compiled is based on the C99 standard
which ought to be very seasoned for at least a decade, i.e., I would
be *very* surprised if compilers introduced major incompatibilites
with this code in the future. IOW, let's just use the default versions
of the host system...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>